% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot bar graph from data frame}
\usage{
plot_bar(x, ax, ay, angle = 0)
}
\arguments{
\item{x}{A data.frame}

\item{ax}{A column name of X axis.}

\item{ay}{A column name of Y axis.}

\item{angle}{A angle of X axis ticks (in degrees)}
}
\value{
The plot of \code{x} with \code{ax} horizontal axis and \code{ay} vertical axis.
}
\description{
Plot bar graph from data frame
}
\examples{
plot_bar(dataframe, "column1","column2")
plot_bar(dataframe, "column1","column2",angle=45)
}
