% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot histogram from factor (vector)}
\usage{
plot_hist(x, xlab, ylab, binwidth = 0.03)
}
\arguments{
\item{x}{A vector of factor}

\item{xlab}{(string with quotes) A name of X axis label.}

\item{ylab}{(string with quotes) A name of Y axis label.}

\item{binwidth}{A numeric value, width of a bin in histogram}
}
\value{
The histogram plot of \code{x}.
}
\description{
Plot histogram from factor (vector)
}
\examples{
plot_hist(vector, "label of x axis","label of x axis")
plot_hist(vector, "label of x axis","label of x axis",binwidth=0.5)
}
